requirements.txt
streamlit
tensorflow
transformers
torch
pillow
scikit-learn
